chore(connector): Deprecate connectors.base
