+++ tests/integration_tests/base_tests.py
@@ -36,7 +36,6 @@ from sqlalchemy.dialects.mysql import dialect
 from tests.integration_tests.test_app import app, login
 from superset.sql_parse import CtasMethod
 from superset import db, security_manager
-from superset.connectors.base.models import BaseDatasource
 from superset.connectors.sqla.models import SqlaTable
 from superset.models import core as models
 from superset.models.slice import Slice
@@ -236,7 +235,7 @@ class SupersetTestCase(TestCase):
             raise ValueError("Database doesn't exist")
 
     @staticmethod
-    def get_datasource_mock() -> BaseDatasource:
+    def get_datasource_mock() -> SqlaTable:
         datasource = MagicMock()
         results = Mock()
         results.query = Mock()
